Design implementation
This implementation is part of the full project.

A url rewrite mod have been enabled for this project for the sake of
friendly url. However, to use this project with your WAMP server you
would need to enable the rewrite_module in the apache module and then
you are good to go. Please remember to run the database sql script in
the project folder...it will help you guyz get started....Salam